midi_pattern_generator/midi_utils.py
```python
import random
from midi_pattern_generator.constants import MIDI_NOTE_RANGE

def midi_to_note(midi_number):
    """
    Converts a MIDI number to a note.
    """
    notes = ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']
    octave = midi_number // 12 - 1
    note = notes[midi_number % 12]
    return note + str(octave)

def note_to_midi(note):
    """
    Converts a note to a MIDI number.
    """
    notes = ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']
    note_name = note[:-1]
    octave = int(note[-1]) + 1
    midi_number = notes.index(note_name) + octave * 12
    return midi_number

def generate_random_note():
    """
    Generates a random note within the MIDI note range.
    """
    return midi_to_note(random.randint(*MIDI_NOTE_RANGE))

def validate_note(note):
    """
    Validates if a note is within the MIDI note range.
    """
    midi_number = note_to_midi(note)
    return MIDI_NOTE_RANGE[0] <= midi_number <= MIDI_NOTE_RANGE[1]
```